Crate human_regex

A Regex for Humans

The goal of this crate is simple: give everybody the power of regular expressions without having to learn the complicated syntax. It is inspired by ReadableRegex.jl.

Example usage

Matching a date

If you want to match a date of the format 2021-10-30, you would use the following code to generate a regex:

let hr = human_regex::HumanRegex::new()
    .begin()
    .exactly(4, human_regex::DIGIT)
    .text("-")
    .exactly(2, human_regex::DIGIT)
    .text("-")
    .exactly(2, human_regex::DIGIT)
    .end();
assert!(hr.is_match("2014-01-01"));

Specifically, this chunk of code would yield the regex ^\d{4}-\d{2}-\d{2}$, which is exactly what we want!

Structs

HumanRegex

The HumanRegex struct which maintains and updates the regex string

Constants

ANY

A constant for matching any character (except for \n)

DIGIT

A constant for the digit character class (i.e., the digits 0 through 9)

NON_DIGIT

A constant for the non-digit character class (i.e., everything BUT the digits 0-9)

NON_WHITESPACE

A constant for the whitespace character class (i.e., everything BUT space and tab)

NON_WORD

A constant for the non-word character class (i.e., everything BUT the alphanumeric characters plus underscore)

WHITESPACE

A constant for the whitespace character class (i.e., space and tab)

WORD

A constant for the word character class (i.e., all alphanumeric characters plus underscore)