Struct human_regex::HumanRegex

pub struct HumanRegex {
    pub regex_string: String,
}

The HumanRegex struct which maintains and updates the regex string

Fields
regex_string: String

The internally-maintained true regex string

Implementations

impl HumanRegex

pub fn new() -> Self

Generate a new HumanRegex with a blank regex_string

pub fn exactly<T>(&self, n: u8, target: T) -> Self where
T: Into<String> + Display,

Match exactly n of a certain target

pub fn at_least(&self, n: u8, target: &str) -> Self

Match at least n of a certain target

pub fn at_least_at_most(&self, n: u8, m: u8, target: &str) -> Self

Match at least n and at most m of a certain target

pub fn one_or_more(&self, target: &str) -> Self

Match one or more of a certain target

pub fn zero_or_more(&self, target: &str) -> Self

Match zero or more of a certain target

pub fn zero_or_one(&self, target: &str) -> Self

Match zero or one of a certain target

pub fn text<T>(&self, text: T) -> Self where
T: Into<String> + Display,

Add text directly to the match string

pub fn begin(&self) -> Self

Represents the beginning of the text

pub fn end(&self) -> Self

Represents the end of the text

pub fn from_regex_string(regex_string: &str) -> Self

Generates a new human regex directly from a regex string

pub fn get_regex(&self) -> Regex

Returns the current state of the constructed regex string as a Regex object

pub fn is_match(&self, string_to_match: &str) -> bool

Checks whether or not a string matches with the constructed regex